+++ elasticapm/transport/base.py
@@ -3,6 +3,7 @@ import gzip
 import logging
 import threading
 import timeit
+from collections import defaultdict
 
 from elasticapm.contrib.async_worker import AsyncWorker
 from elasticapm.utils import json_encoder
@@ -48,19 +49,23 @@ class Transport(object):
         """
         self.state = TransportState()
         self._metadata = metadata if metadata is not None else {}
-        self._compress_level = compress_level
+        self._compress_level = min(9, max(0, compress_level if compress_level is not None else 0))
         self._json_serializer = json_serializer
         self._max_flush_time = max_flush_time
         self._max_buffer_size = max_buffer_size
         self._queued_data = None
-        self._flush_lock = threading.Lock()
+        self._queue_lock = threading.Lock()
         self._last_flush = timeit.default_timer()
         self._flush_timer = None
+        self._counts = defaultdict(int)
 
     def queue(self, event_type, data, flush=False):
-        self._queue(self.queued_data, {event_type: data})
-        since_last_flush = timeit.default_timer() - self._last_flush
-        queue_size = self.queued_data_size
+        with self._queue_lock:
+            queued_data = self.queued_data
+            queued_data.write((self._json_serializer({event_type: data}) + "\n").encode("utf-8"))
+            self._counts[event_type] += 1
+            since_last_flush = timeit.default_timer() - self._last_flush
+            queue_size = 0 if queued_data.fileobj is None else queued_data.fileobj.tell()
         if flush:
             logger.debug("forced flush")
             self.flush()
@@ -77,27 +82,16 @@ class Transport(object):
             )
             self.flush()
         elif not self._flush_timer:
-            with self._flush_lock:
+            with self._queue_lock:
                 self._start_flush_timer()
-
-    def _queue(self, queue, data):
-        queue.write((self._json_serializer(data) + "\n").encode("utf-8"))
 
     @property
     def queued_data(self):
         if self._queued_data is None:
-            if self._compress_level:
-                self._queued_data = gzip.GzipFile(fileobj=BytesIO(), mode="w", compresslevel=self._compress_level)
-            else:
-                self._queued_data = BytesIO()
-            self._queue(self._queued_data, {"metadata": self._metadata})
+            self._queued_data = gzip.GzipFile(fileobj=BytesIO(), mode="w", compresslevel=self._compress_level)
+            data = (self._json_serializer({"metadata": self._metadata}) + "\n").encode("utf-8")
+            self._queued_data.write(data)
         return self._queued_data
-
-    @property
-    def queued_data_size(self):
-        f = self.queued_data
-        # return size of the underlying BytesIO object if it is compressed
-        return f.fileobj.tell() if hasattr(f, "fileobj") else f.tell()
 
     def flush(self, sync=False, start_flush_timer=True):
         """
@@ -106,17 +100,15 @@ class Transport(object):
         :param start_flush_timer: set to True if the flush timer thread should be restarted at the end of the flush
         :return: None
         """
-        with self._flush_lock:
+        with self._queue_lock:
             self._stop_flush_timer()
             queued_data, self._queued_data = self._queued_data, None
             if queued_data and not self.state.should_try():
                 logger.error("dropping flushed data due to transport failure back-off")
             elif queued_data:
-                if self._compress_level:
-                    fileobj = queued_data.fileobj  # get a reference to the fileobj before closing the gzip file
-                    queued_data.close()
-                else:
-                    fileobj = queued_data
+                fileobj = queued_data.fileobj  # get a reference to the fileobj before closing the gzip file
+                queued_data.close()
+
                 # StringIO on Python 2 does not have getbuffer, so we need to fall back to getvalue
                 data = fileobj.getbuffer() if hasattr(fileobj, "getbuffer") else fileobj.getvalue()
                 if hasattr(self, "send_async") and not sync: