+++ tests/contrib/flask/flask_tests.py
@@ -2,12 +2,14 @@ import pytest  # isort:skip
 
 pytest.importorskip("flask")  # isort:skip
 
+import logging
 import os
 
 from elasticapm.conf import constants
 from elasticapm.conf.constants import ERROR, TRANSACTION
 from elasticapm.contrib.flask import ElasticAPM
 from elasticapm.utils import compat
+from tests.contrib.flask.utils import captured_templates
 
 try:
     from urllib.request import urlopen
@@ -216,7 +218,7 @@ def test_non_standard_http_status(flask_apm_client):
 
 
 def test_framework_name(flask_app):
-    elasticapm = ElasticAPM(app=flask_app)
+    elasticapm = ElasticAPM(app=flask_app, metrics_interval="0ms")
     assert elasticapm.client.config.framework_name == "flask"
     app_info = elasticapm.client.get_service_info()
     assert app_info["framework"]["name"] == "flask"
@@ -284,3 +286,41 @@ def test_set_transaction_name(flask_apm_client):
     transaction = flask_apm_client.client.events[TRANSACTION][0]
     assert transaction["name"] == "foo"
     assert transaction["result"] == "okydoky"
+
+
+def test_rum_tracing_context_processor(flask_apm_client):
+    with captured_templates(flask_apm_client.app) as templates:
+        resp = flask_apm_client.app.test_client().post("/users/", data={"foo": "bar"})
+        resp.close()
+        transaction = flask_apm_client.client.events[TRANSACTION][0]
+        template, context = templates[0]
+        assert context["apm"]["trace_id"] == transaction["trace_id"]
+        assert context["apm"]["is_sampled"]
+        assert context["apm"]["is_sampled_js"] == "true"
+        assert callable(context["apm"]["span_id"])
+
+
+@pytest.mark.parametrize("flask_apm_client", [{"logging": True}], indirect=True)
+def test_logging_enabled(flask_apm_client):
+    logger = logging.getLogger()
+    logger.error("test")
+    error = flask_apm_client.client.events[ERROR][0]
+    assert error["log"]["level"] == "error"
+    assert error["log"]["message"] == "test"
+
+
+@pytest.mark.parametrize("flask_apm_client", [{"logging": False}], indirect=True)
+def test_logging_disabled(flask_apm_client):
+    logger = logging.getLogger()
+    logger.error("test")
+    assert len(flask_apm_client.client.events[ERROR]) == 0
+
+
+@pytest.mark.parametrize("flask_apm_client", [{"logging": logging.ERROR}], indirect=True)
+def test_logging_by_level(flask_apm_client):
+    logger = logging.getLogger()
+    logger.warning("test")
+    logger.error("test")
+    assert len(flask_apm_client.client.events[ERROR]) == 1
+    error = flask_apm_client.client.events[ERROR][0]
+    assert error["log"]["level"] == "error"