+++ tests/instrumentation/urllib3_tests.py
@@ -1,4 +1,4 @@
-import mock
+import pytest
 import urllib3
 
 from elasticapm.conf import constants
@@ -14,7 +14,7 @@ def test_urllib3(instrument, elasticapm_client, waiting_httpserver):
     parsed_url = urlparse.urlparse(url)
     elasticapm_client.begin_transaction("transaction")
     expected_sig = "GET {0}".format(parsed_url.netloc)
-    with capture_span("test_pipeline", "test"):
+    with capture_span("test_name", "test_type"):
         pool = urllib3.PoolManager(timeout=0.1)
 
         url = "http://{0}/hello_world".format(parsed_url.netloc)
@@ -25,7 +25,7 @@ def test_urllib3(instrument, elasticapm_client, waiting_httpserver):
     transactions = elasticapm_client.events[TRANSACTION]
     spans = elasticapm_client.spans_for_transaction(transactions[0])
 
-    expected_signatures = {"test_pipeline", expected_sig}
+    expected_signatures = {"test_name", expected_sig}
 
     assert {t["name"] for t in spans} == expected_signatures
 
@@ -36,8 +36,8 @@ def test_urllib3(instrument, elasticapm_client, waiting_httpserver):
     assert spans[0]["context"]["url"] == url
     assert spans[0]["parent_id"] == spans[1]["id"]
 
-    assert spans[1]["name"] == "test_pipeline"
-    assert spans[1]["type"] == "test"
+    assert spans[1]["name"] == "test_name"
+    assert spans[1]["type"] == "test_type"
     assert spans[1]["parent_id"] == transactions[0]["id"]
 
 
@@ -76,3 +76,22 @@ def test_trace_parent_propagation_unsampled(instrument, elasticapm_client, waiti
     assert trace_parent.trace_id == transactions[0]["trace_id"]
     assert trace_parent.span_id == transaction_object.id
     assert not trace_parent.trace_options.recorded
+
+
+@pytest.mark.parametrize("elasticapm_client", [{"transaction_max_spans": 1}], indirect=True)
+def test_span_only_dropped(instrument, elasticapm_client, waiting_httpserver):
+    """test that urllib3 instrumentation does not fail if no parent span can be found"""
+    waiting_httpserver.serve_content("")
+    url = waiting_httpserver.url + "/hello_world"
+    transaction_object = elasticapm_client.begin_transaction("transaction")
+    for i in range(2):
+        with capture_span("test", "test"):
+            pool = urllib3.PoolManager(timeout=0.1)
+            pool.request("GET", url)
+    elasticapm_client.end_transaction("bla", "OK")
+    trace_parent_1 = TraceParent.from_string(waiting_httpserver.requests[0].headers[constants.TRACEPARENT_HEADER_NAME])
+    trace_parent_2 = TraceParent.from_string(waiting_httpserver.requests[1].headers[constants.TRACEPARENT_HEADER_NAME])
+
+    assert trace_parent_1.span_id != transaction_object.id
+    # second request should use transaction id as span id because there is no span
+    assert trace_parent_2.span_id == transaction_object.id