+++ CHANGELOG.md
@@ -1,9 +1,36 @@
 # Changelog
 
+## Unreleased
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.0.3...master)
+
+ * Added support for collecting system and process metrics (#361)
+ * Added `transaction.sampled` to errors (#371)
+ * Added parsing of `/proc/self/cgroup` to capture container meta data (#352)
+ * Added option to configure logging for Flask using a log level (#344)
+
+## v4.0.3
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.0.2...v4.0.3)
+
+ * implemented de-dotting of tag names and context keys (#353)
+ * wrote a quickfix for the boto3/botocore instrumentation (#367)
+ * fixed an issue with psycopg2 and encoded strings (#366)
+
+## v4.0.2
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.0.1...v4.0.2)
+
+ * fixed another issue in the new v2 transport (#351)
+
+## v4.0.1
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.0.0...v4.0.1)
+
+ * fixed an issue with instrumenting redis-py 3.0+
+ * fixed a multithreading issue that occurs when using threaded workers (#335)
+ 
 ## v4.0.0
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v3.0.2...v4.0.0)
 
 **BREAKING** Version 4 of the agent implements a new wire protocol for communicating with
-the APM Server. This format is only supported in APM Server 6.5+.
+the APM Server. This format is only supported in *APM Server 6.5+*.
 
 Further breaking changes: